Add files via upload
Pixscribe App
Pixscribe is an AI-powered application built with Streamlit that transforms images into meaningful text. Whether you need to extract text from an image, get detailed scene descriptions, or listen to the extracted text, Pixscribe has you covered. It leverages the Google Gemini API for scene understanding, Tesseract OCR for text extraction, and pyttsx3 for text-to-speech functionality. Designed with accessibility in mind, Pixscribe makes it easier for users, including the visually impaired, to interact with visual content in a whole new way.